Added ability to add images (W.I.P. in terms of proper folder path etc.)

+++ src/main/java/zti_spring_backend/rest_api/OfferController.java
@@ -167,6 +167,9 @@ public class OfferController {
         if(user.getId() != offer.getSellerid())
             throw new OfferNotFoundException(id, user.getId());
 
+        if(orderRepository.countByOfferid(id) > 0)
+            throw new OfferInvalidDeleteOperationException("Cannot delete an offer with orders.");
+
         offerRepository.deleteById(id);
         return ResponseEntity.ok(null);
     }
